R!! Replaces 'webpackTarget' and 'profiles' with new 'profile' property.
Breaking: You can now use only a single property with 'profile' that should result in the same output as use of the prior 'webpackTarget' property. If you have used multiple profiles as a chain of profiles, try the new 'depends' property in the CompilationProfile configuration.

Introducing Profiles: Replace compilation targets with profiles #55